pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "storage-drift"
version = "0.1.0"
description = "Slither plugin that detects storage slots that diverge across chains — values dependent on gas mechanics, CREATE2, or sender balance that differ between execution environments (e.g. Ethereum vs Polkadot/revive)."
license = "AGPL-3.0-only"
requires-python = ">=3.9"
dependencies = ["slither-analyzer>=0.6.0"]

[project.optional-dependencies]
dev = ["pytest", "ruff", "solc-select"]

[project.entry-points."slither_analyzer.plugin"]
storage_drift = "storage_drift:make_plugin"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py39"
line-length = 79

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
